Work on Debugger
- plug Debugger with WireProtocol in plain C
- fix #226
- completly replaced WireProtocol C++ with C version
- replace private_malloc with platform_malloc
- rework Monitor_FlashSectorMap to report full details on BlockStorage
- fixes #228
- further improvements in BlockStorage
- add stubs for COM
- add interface from BlockStorage interface to STM32 Flash driver
- rework nanoCLR VS Solution to make it compatible with latest changes
- revert several types to well know types
- add weak declarations to COM stubs in Win32
- correct linker scripts to have addresses aligned with sector boundaries
- add block storage definitions for all reference boards
- delete CMake file for NF_WireProtocol (is already in NF_Debugger)

+++ targets/os/win32/nanoCLR/PAL/time_stubs.cpp
@@ -45,7 +45,7 @@ signed __int64 Time_GetTickCount()
     return 0;
 }
 
-bool HAL_Time_ToSystemTime(signed __int64 time, SYSTEMTIME* systemTime)
+bool HAL_Time_ToSystemTime(signed long long time, SYSTEMTIME* systemTime)
 {
     return FALSE;
 }
